// web/src/api/generated/models/MilestoneDetail.ts
export interface MilestoneDetail {
    /**
     * 
     * @type {AuthorSummary}
     * @memberof MilestoneDetail
     */
    author: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDetail
     */
    authorId: string;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDetail
     */
    body: string | null;
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof MilestoneDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneDetail
     */
    happenedOn: Date;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDetail
     */
    spaceId: string;
    /**
     * 
     * @type {string}
     * @memberof MilestoneDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof MilestoneDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof MilestoneDetail
     */
    version: number;
}